Added code to MainActivity.java and arrays.xml
Set up MainActivity.java compileDirection() method to work with location
in array (to determine if the user will go left or right).  Added a LOT
to the array file so that we are including EVERYTHING (stairs,
bathrooms, etc) for CC1 and CC2.   Also added our three special comments
"room is behind student breakout area" or "room is with a cluster of
rooms off of the main hall" or "room is close to the large window across
from the stairs" which take care of our special rooms.  I have NOT yet
added these responses into the java file (that will be next - I can do
that on Monday). - jenB